Split `GradleEnterpriseExtensionApplicationTest` to better utilize parallel testing.
