+++ agent/src/test/groovy/nu/studer/teamcity/buildscan/agent/maven/GradleEnterpriseExtensionApplicationTest.groovy
@@ -193,420 +193,4 @@ class GradleEnterpriseExtensionApplicationTest extends BaseExtensionApplicationT
         jdkCompatibleMavenVersion << SUPPORTED_MAVEN_VERSIONS
     }
 
-    def "does not inject GE extension when not defined in project but matching custom coordinates defined in project (#jdkCompatibleMavenVersion)"() {
-        assumeTrue jdkCompatibleMavenVersion.isJvmVersionCompatible()
-        assumeTrue GE_URL != null
-
-        given:
-        def mvnProject = new MavenProject.Configuration(
-            geUrl: GE_URL,
-            // using Guava as surrogate since we do not have a custom extension at hand that pulls in the GE Maven extension transitively
-            customExtension: new GroupArtifactVersion(group: 'com.google.guava', artifact: 'guava', version: '31.1-jre')
-        ).buildIn(checkoutDir)
-
-        and:
-        def gePluginConfig = new TcPluginConfig(
-            geUrl: GE_URL,
-            geExtensionVersion: GE_EXTENSION_VERSION,
-            geExtensionCustomCoordinates: 'com.google.guava:guava',
-        )
-
-        when:
-        def output = run(jdkCompatibleMavenVersion.mavenVersion, mvnProject, gePluginConfig)
-
-        then:
-        0 * extensionApplicationListener.geExtensionApplied(_)
-        0 * extensionApplicationListener.ccudExtensionApplied(_)
-
-        and:
-        outputMissesTeamCityServiceMessageBuildStarted(output)
-        outputMissesTeamCityServiceMessageBuildScanUrl(output)
-
-        where:
-        jdkCompatibleMavenVersion << SUPPORTED_MAVEN_VERSIONS
-    }
-
-    def "applies CCUD extension via classpath when not defined in project where GE extension not defined in project and not applied via classpath (#jdkCompatibleMavenVersion)"() {
-        assumeTrue jdkCompatibleMavenVersion.isJvmVersionCompatible()
-        assumeTrue GE_URL != null
-
-        given:
-        def mvnProject = new MavenProject.Configuration().buildIn(checkoutDir)
-
-        and:
-        def gePluginConfig = new TcPluginConfig(
-            geUrl: GE_URL,
-            ccudExtensionVersion: CCUD_EXTENSION_VERSION,
-        )
-
-        when:
-        def output = run(jdkCompatibleMavenVersion.mavenVersion, mvnProject, gePluginConfig)
-
-        then:
-        0 * extensionApplicationListener.geExtensionApplied(_)
-        1 * extensionApplicationListener.ccudExtensionApplied(CCUD_EXTENSION_VERSION)
-
-        and:
-        outputMissesTeamCityServiceMessageBuildStarted(output)
-        outputMissesTeamCityServiceMessageBuildScanUrl(output)
-
-        where:
-        jdkCompatibleMavenVersion << SUPPORTED_MAVEN_VERSIONS
-    }
-
-    def "applies CCUD extension via classpath when not defined in project where GE extension applied via classpath (#jdkCompatibleMavenVersion)"() {
-        assumeTrue jdkCompatibleMavenVersion.isJvmVersionCompatible()
-        assumeTrue GE_URL != null
-
-        given:
-        def mvnProject = new MavenProject.Configuration().buildIn(checkoutDir)
-
-        and:
-        def gePluginConfig = new TcPluginConfig(
-            geUrl: GE_URL,
-            geExtensionVersion: GE_EXTENSION_VERSION,
-            ccudExtensionVersion: CCUD_EXTENSION_VERSION,
-        )
-
-        when:
-        def output = run(jdkCompatibleMavenVersion.mavenVersion, mvnProject, gePluginConfig)
-
-        then:
-        1 * extensionApplicationListener.geExtensionApplied(GE_EXTENSION_VERSION)
-        1 * extensionApplicationListener.ccudExtensionApplied(CCUD_EXTENSION_VERSION)
-
-        and:
-        outputContainsTeamCityServiceMessageBuildStarted(output)
-        outputContainsTeamCityServiceMessageBuildScanUrl(output)
-
-        where:
-        jdkCompatibleMavenVersion << SUPPORTED_MAVEN_VERSIONS
-    }
-
-    def "applies CCUD extension via classpath when not defined in project where GE extension defined in project (#jdkCompatibleMavenVersion)"() {
-        assumeTrue jdkCompatibleMavenVersion.isJvmVersionCompatible()
-        assumeTrue GE_URL != null
-
-        given:
-        def mvnProject = new MavenProject.Configuration(
-            geUrl: GE_URL,
-            geExtensionVersion: GE_EXTENSION_VERSION,
-        ).buildIn(checkoutDir)
-
-        and:
-        def gePluginConfig = new TcPluginConfig(
-            geUrl: GE_URL,
-            geExtensionVersion: GE_EXTENSION_VERSION,
-            ccudExtensionVersion: CCUD_EXTENSION_VERSION,
-        )
-
-        when:
-        def output = run(jdkCompatibleMavenVersion.mavenVersion, mvnProject, gePluginConfig)
-
-        then:
-        0 * extensionApplicationListener.geExtensionApplied(_)
-        1 * extensionApplicationListener.ccudExtensionApplied(CCUD_EXTENSION_VERSION)
-
-        and:
-        outputContainsTeamCityServiceMessageBuildStarted(output)
-        outputContainsTeamCityServiceMessageBuildScanUrl(output)
-
-        where:
-        jdkCompatibleMavenVersion << SUPPORTED_MAVEN_VERSIONS
-    }
-
-    def "applies CCUD extension via project when defined in project (#jdkCompatibleMavenVersion)"() {
-        assumeTrue jdkCompatibleMavenVersion.isJvmVersionCompatible()
-        assumeTrue GE_URL != null
-
-        given:
-        def mvnProject = new MavenProject.Configuration(
-            geUrl: GE_URL,
-            geExtensionVersion: GE_EXTENSION_VERSION,
-            ccudExtensionVersion: CCUD_EXTENSION_VERSION
-        ).buildIn(checkoutDir)
-
-        and:
-        def gePluginConfig = new TcPluginConfig(
-            geUrl: GE_URL,
-            geExtensionVersion: GE_EXTENSION_VERSION,
-            ccudExtensionVersion: CCUD_EXTENSION_VERSION,
-        )
-
-        when:
-        def output = run(jdkCompatibleMavenVersion.mavenVersion, mvnProject, gePluginConfig)
-
-        then:
-        0 * extensionApplicationListener.geExtensionApplied(_)
-        0 * extensionApplicationListener.ccudExtensionApplied(_)
-
-        and:
-        outputContainsTeamCityServiceMessageBuildStarted(output)
-        outputContainsTeamCityServiceMessageBuildScanUrl(output)
-
-        where:
-        jdkCompatibleMavenVersion << SUPPORTED_MAVEN_VERSIONS
-    }
-
-    def "does not inject CCUD extension when not defined in project but matching custom coordinates defined in project (#jdkCompatibleMavenVersion)"() {
-        assumeTrue jdkCompatibleMavenVersion.isJvmVersionCompatible()
-        assumeTrue GE_URL != null
-
-        given:
-        def mvnProject = new MavenProject.Configuration(
-            geUrl: GE_URL,
-            geExtensionVersion: GE_EXTENSION_VERSION,
-            // using Guava as surrogate since we do not have a custom extension at hand that pulls in the GE Maven extension transitively
-            customExtension: new GroupArtifactVersion(group: 'com.google.guava', artifact: 'guava', version: '31.1-jre')
-        ).buildIn(checkoutDir)
-
-        and:
-        def gePluginConfig = new TcPluginConfig(
-            geUrl: GE_URL,
-            ccudExtensionVersion: CCUD_EXTENSION_VERSION,
-            ccudExtensionCustomCoordinates: 'com.google.guava:guava',
-        )
-
-        when:
-        def output = run(jdkCompatibleMavenVersion.mavenVersion, mvnProject, gePluginConfig)
-
-        then:
-        0 * extensionApplicationListener.geExtensionApplied(_)
-        0 * extensionApplicationListener.ccudExtensionApplied(_)
-
-        and:
-        outputContainsTeamCityServiceMessageBuildStarted(output)
-        outputContainsTeamCityServiceMessageBuildScanUrl(output)
-
-        where:
-        jdkCompatibleMavenVersion << SUPPORTED_MAVEN_VERSIONS
-    }
-
-    def "ignores GE URL requested via TC config when GE extension is not applied via the classpath (#jdkCompatibleMavenVersion)"() {
-        assumeTrue jdkCompatibleMavenVersion.isJvmVersionCompatible()
-        assumeTrue GE_URL != null
-
-        given:
-        def mvnProject = new MavenProject.Configuration(
-            geUrl: GE_URL,
-            geExtensionVersion: GE_EXTENSION_VERSION,
-        ).buildIn(checkoutDir)
-
-        and:
-        def gePluginConfig = new TcPluginConfig(
-            geUrl: new URI('https://ge-server.invalid'),
-            geAllowUntrustedServer: true,
-            geExtensionVersion: GE_EXTENSION_VERSION,
-        )
-
-        when:
-        def output = run(jdkCompatibleMavenVersion.mavenVersion, mvnProject, gePluginConfig)
-
-        then:
-        0 * extensionApplicationListener.geExtensionApplied(_)
-        0 * extensionApplicationListener.ccudExtensionApplied(_)
-
-        and:
-        outputContainsTeamCityServiceMessageBuildStarted(output)
-        outputContainsTeamCityServiceMessageBuildScanUrl(output)
-
-        where:
-        jdkCompatibleMavenVersion << SUPPORTED_MAVEN_VERSIONS
-    }
-
-    def "configures GE URL requested via TC config when GE extension is applied via classpath (#jdkCompatibleMavenVersion)"() {
-        assumeTrue jdkCompatibleMavenVersion.isJvmVersionCompatible()
-        assumeTrue GE_URL != null
-
-        given:
-        def mvnProject = new MavenProject.Configuration(
-            geUrl: 'https://ge-server.invalid',
-            geExtensionVersion: null,
-        ).buildIn(checkoutDir)
-
-        and:
-        def gePluginConfig = new TcPluginConfig(
-            geUrl: GE_URL,
-            geAllowUntrustedServer: true,
-            geExtensionVersion: GE_EXTENSION_VERSION,
-        )
-
-        when:
-        def output = run(jdkCompatibleMavenVersion.mavenVersion, mvnProject, gePluginConfig)
-
-        then:
-        1 * extensionApplicationListener.geExtensionApplied(GE_EXTENSION_VERSION)
-        0 * extensionApplicationListener.ccudExtensionApplied(_)
-
-        and:
-        outputContainsTeamCityServiceMessageBuildStarted(output)
-        outputContainsTeamCityServiceMessageBuildScanUrl(output)
-
-        where:
-        jdkCompatibleMavenVersion << SUPPORTED_MAVEN_VERSIONS
-    }
-
-    def "does not publish build scan for TeamCity specific info goal invocation (#jdkCompatibleMavenVersion)"() {
-        assumeTrue jdkCompatibleMavenVersion.isJvmVersionCompatible()
-        assumeTrue GE_URL != null
-
-        given:
-        def mvnProject = new MavenProject.Configuration().buildIn(checkoutDir)
-
-        and:
-        def gePluginConfig = new TcPluginConfig(
-            geUrl: GE_URL,
-            geExtensionVersion: GE_EXTENSION_VERSION,
-        )
-
-        and:
-        def mvnBuildStepConfig = new MavenBuildStepConfig(
-            checkoutDir: checkoutDir,
-            goals: 'org.jetbrains.maven:info-maven3-plugin:1.0.2:info',
-        )
-
-        when:
-        def output = run(jdkCompatibleMavenVersion.mavenVersion, mvnProject, gePluginConfig, mvnBuildStepConfig)
-
-        then:
-        outputMissesTeamCityServiceMessageBuildStarted(output)
-        outputMissesTeamCityServiceMessageBuildScanUrl(output)
-
-        where:
-        jdkCompatibleMavenVersion << SUPPORTED_MAVEN_VERSIONS
-    }
-
-    def "build succeeds when service message maven extension is applied to a project without GE in the extension classpath (#jdkCompatibleMavenVersion)"() {
-        assumeTrue jdkCompatibleMavenVersion.isJvmVersionCompatible()
-        assumeTrue GE_URL != null
-
-        given:
-        def mvnProject = new MavenProject.Configuration().buildIn(checkoutDir)
-
-        and:
-        def gePluginConfig = new TcPluginConfig(
-            enableCommandLineRunner: true,
-        )
-
-        when:
-        def output = run(jdkCompatibleMavenVersion.mavenVersion, mvnProject, gePluginConfig)
-
-        then:
-        outputContainsBuildSuccess(output)
-        outputMissesTeamCityServiceMessageBuildStarted(output)
-        outputMissesTeamCityServiceMessageBuildScanUrl(output)
-
-        where:
-        jdkCompatibleMavenVersion << SUPPORTED_MAVEN_VERSIONS
-    }
-
-    def "publishes build scan when pom is in a subdirectory and extensions.xml is in project root directory (#jdkCompatibleMavenVersion)"() {
-        assumeTrue jdkCompatibleMavenVersion.isJvmVersionCompatible()
-        assumeTrue GE_URL != null
-
-        given:
-        def mvnProject = new MavenProject.Configuration(
-            geUrl: GE_URL,
-            geExtensionVersion: GE_EXTENSION_VERSION,
-            pomDirName: 'subdir'
-        ).buildIn(checkoutDir)
-
-        and:
-        def gePluginConfig = new TcPluginConfig(
-            geExtensionVersion: GE_EXTENSION_VERSION,
-        )
-
-        and:
-        def mvnBuildStepConfig = new MavenBuildStepConfig(
-            checkoutDir: checkoutDir,
-            pathToPomFile: getRelativePath(checkoutDir, mvnProject.pom)
-        )
-
-        when:
-        def output = run(jdkCompatibleMavenVersion.mavenVersion, mvnProject, gePluginConfig, mvnBuildStepConfig)
-
-        then:
-        outputContainsBuildSuccess(output)
-
-        and:
-        outputContainsTeamCityServiceMessageBuildStarted(output)
-        outputContainsTeamCityServiceMessageBuildScanUrl(output)
-
-        where:
-        jdkCompatibleMavenVersion << SUPPORTED_MAVEN_VERSIONS
-    }
-
-    def "publishes build scan when pom is in a subdirectory and subdirectory is specified as pom path and extensions.xml is in project root directory (#jdkCompatibleMavenVersion)"() {
-        assumeTrue jdkCompatibleMavenVersion.isJvmVersionCompatible()
-        assumeTrue GE_URL != null
-
-        given:
-        def mvnProject = new MavenProject.Configuration(
-            geUrl: GE_URL,
-            geExtensionVersion: GE_EXTENSION_VERSION,
-            pomDirName: 'subdir'
-        ).buildIn(checkoutDir)
-
-        and:
-        def gePluginConfig = new TcPluginConfig(
-            geExtensionVersion: GE_EXTENSION_VERSION,
-        )
-
-        and:
-        def mvnBuildStepConfig = new MavenBuildStepConfig(
-            checkoutDir: checkoutDir,
-            pathToPomFile: getRelativePath(checkoutDir, mvnProject.pom.parentFile)
-        )
-
-        when:
-        def output = run(jdkCompatibleMavenVersion.mavenVersion, mvnProject, gePluginConfig, mvnBuildStepConfig)
-
-        then:
-        outputContainsBuildSuccess(output)
-
-        and:
-        outputContainsTeamCityServiceMessageBuildStarted(output)
-        outputContainsTeamCityServiceMessageBuildScanUrl(output)
-
-        where:
-        jdkCompatibleMavenVersion << SUPPORTED_MAVEN_VERSIONS
-    }
-
-    def "publishes build scan when pom is in a subdirectory and extensions.xml is in a higher subdirectory (#jdkCompatibleMavenVersion)"() {
-        assumeTrue jdkCompatibleMavenVersion.isJvmVersionCompatible()
-        assumeTrue GE_URL != null
-
-        given:
-        def mvnProject = new MavenProject.Configuration(
-            geUrl: GE_URL,
-            geExtensionVersion: GE_EXTENSION_VERSION,
-            pomDirName: 'subdir1/subdir2',
-            dotMvnParentDirName: 'subdir1',
-        ).buildIn(checkoutDir)
-
-        and:
-        def gePluginConfig = new TcPluginConfig(
-            geExtensionVersion: GE_EXTENSION_VERSION,
-        )
-
-        and:
-        def mvnBuildStepConfig = new MavenBuildStepConfig(
-            checkoutDir: checkoutDir,
-            pathToPomFile: getRelativePath(checkoutDir, mvnProject.pom)
-        )
-
-        when:
-        def output = run(jdkCompatibleMavenVersion.mavenVersion, mvnProject, gePluginConfig, mvnBuildStepConfig)
-
-        then:
-        outputContainsBuildSuccess(output)
-
-        and:
-        outputContainsTeamCityServiceMessageBuildStarted(output)
-        outputContainsTeamCityServiceMessageBuildScanUrl(output)
-
-        where:
-        jdkCompatibleMavenVersion << SUPPORTED_MAVEN_VERSIONS
-    }
-
 }
