+++ agent/src/test/groovy/nu/studer/teamcity/buildscan/agent/maven/GradleEnterpriseUrlConfigurationTest.groovy
@@ -1,0 +1,73 @@
+package nu.studer.teamcity.buildscan.agent.maven
+
+import nu.studer.teamcity.buildscan.agent.TcPluginConfig
+
+import static org.junit.Assume.assumeTrue
+
+class GradleEnterpriseUrlConfigurationTest extends BaseExtensionApplicationTest {
+
+    def "ignores GE URL requested via TC config when GE extension is not applied via the classpath (#jdkCompatibleMavenVersion)"() {
+        assumeTrue jdkCompatibleMavenVersion.isJvmVersionCompatible()
+        assumeTrue GE_URL != null
+
+        given:
+        def mvnProject = new MavenProject.Configuration(
+            geUrl: GE_URL,
+            geExtensionVersion: GE_EXTENSION_VERSION,
+        ).buildIn(checkoutDir)
+
+        and:
+        def gePluginConfig = new TcPluginConfig(
+            geUrl: new URI('https://ge-server.invalid'),
+            geAllowUntrustedServer: true,
+            geExtensionVersion: GE_EXTENSION_VERSION,
+        )
+
+        when:
+        def output = run(jdkCompatibleMavenVersion.mavenVersion, mvnProject, gePluginConfig)
+
+        then:
+        0 * extensionApplicationListener.geExtensionApplied(_)
+        0 * extensionApplicationListener.ccudExtensionApplied(_)
+
+        and:
+        outputContainsTeamCityServiceMessageBuildStarted(output)
+        outputContainsTeamCityServiceMessageBuildScanUrl(output)
+
+        where:
+        jdkCompatibleMavenVersion << SUPPORTED_MAVEN_VERSIONS
+    }
+
+    def "configures GE URL requested via TC config when GE extension is applied via classpath (#jdkCompatibleMavenVersion)"() {
+        assumeTrue jdkCompatibleMavenVersion.isJvmVersionCompatible()
+        assumeTrue GE_URL != null
+
+        given:
+        def mvnProject = new MavenProject.Configuration(
+            geUrl: 'https://ge-server.invalid',
+            geExtensionVersion: null,
+        ).buildIn(checkoutDir)
+
+        and:
+        def gePluginConfig = new TcPluginConfig(
+            geUrl: GE_URL,
+            geAllowUntrustedServer: true,
+            geExtensionVersion: GE_EXTENSION_VERSION,
+        )
+
+        when:
+        def output = run(jdkCompatibleMavenVersion.mavenVersion, mvnProject, gePluginConfig)
+
+        then:
+        1 * extensionApplicationListener.geExtensionApplied(GE_EXTENSION_VERSION)
+        0 * extensionApplicationListener.ccudExtensionApplied(_)
+
+        and:
+        outputContainsTeamCityServiceMessageBuildStarted(output)
+        outputContainsTeamCityServiceMessageBuildScanUrl(output)
+
+        where:
+        jdkCompatibleMavenVersion << SUPPORTED_MAVEN_VERSIONS
+    }
+
+}
